IMS-VT: Modification of data usage aidl
Change data usage aidl interface to take parameter type long
instead of int

CRs-Fixed: 747694

+++ telecomm/java/android/telecom/Connection.java
@@ -447,7 +447,7 @@ public abstract class Connection {
          *
          * @param dataUsage The updated data usage.
          */
-        public void changeCallDataUsage(int dataUsage) {
+        public void changeCallDataUsage(long dataUsage) {
             if (mVideoCallback != null) {
                 try {
                     mVideoCallback.changeCallDataUsage(dataUsage);
